feat(uv/server.cpp): handle server shutdown, allow socket activation
This commit adds:

- socket activation support, to allow the caller to pass a
  pre-existing bound tcp socket to server::run(). This
  is important in contexts where it is necessary to set
  some particular flags on the socket (such as keepalive),
  or if the socket is already created at program start
  (as is the case with the systemd socket activation protocol).
- server shutdown support, to allow signaling the server
  via a std::stop_token that it should clean up and cleanly exit.

Two unit tests were added. Running valgrind on the resulting
binary also revealed a leak due to failure to cleanup a
the uv_handle_t with the connection, this is now taken care of
in the destructor of grpcxx::uv::server::loop_t. Valgrind
does not report any more leaks during test execution.