feat(uv/server): support integration with other event loops
This change adds the ability to subclass grpcxx::uv::server
to be able to trigger individual loop iterations. This makes it
possible to run the server as a nested loop inside either
libuv, or optionally via a different framework (like
libevent).

Additionally, the following two bugs were fixed:

- also support binding to IPv6 addresses
- fix generation of source files from proto files when
  no package is specified.

Note that, due to the current lack of a client-side gRPC
implementation, the tests are not checking the full flow.
Therefore, my approach was to add a sleep in the code and
use Postman to check the service used in the tests answers
correctly.