Handle UV server shutdown and allow socket activation (#42)
* chore(build): add the GTest framework for testing

* feat(uv/server.cpp): handle server shutdown, allow socket activation

This commit adds:

- socket activation support, to allow the caller to pass a
  pre-existing bound tcp socket to server::run(). This
  is important in contexts where it is necessary to set
  some particular flags on the socket (such as keepalive),
  or if the socket is already created at program start
  (as is the case with the systemd socket activation protocol).
- server shutdown support, to allow signaling the server
  via a std::stop_token that it should clean up and cleanly exit.

Two unit tests were added. Running valgrind on the resulting
binary also revealed a leak due to failure to cleanup a
the uv_handle_t with the connection, this is now taken care of
in the destructor of grpcxx::uv::server::loop_t. Valgrind
does not report any more leaks during test execution.

* chore(ci): run tests

* feat(uv/server): support integration with other event loops

This change adds the ability to subclass grpcxx::uv::server
to be able to trigger individual loop iterations. This makes it
possible to run the server as a nested loop inside either
libuv, or optionally via a different framework (like
libevent).

Additionally, the following two bugs were fixed:

- also support binding to IPv6 addresses
- fix generation of source files from proto files when
  no package is specified.

Note that, due to the current lack of a client-side gRPC
implementation, the tests are not checking the full flow.
Therefore, my approach was to add a sleep in the code and
use Postman to check the service used in the tests answers
correctly.

* doc(api): add documentation for new uv/server.h methods

* Update docs/api.md

---------

Co-authored-by: Uditha Atukorala <2161935+uatuko@users.noreply.github.com>