(cmake) update dependencies
Although googletest recommends using the latest commit from the main branch
as often as possible, we also need to know which exact version we are
using for debugging purposes.

We could use a git commit hash for the version but this isn't very human
friendly, so using a tagged version instead.